// src/lib/api/schemas/createPromotionRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationMethod } from './applicationMethod.ts';
import type { PromotionRule } from './promotionRule.ts';
import type { PromotionStatus } from './promotionStatus.ts';
import type { Stacking } from './stacking.ts';

export interface CreatePromotionRequest {
  application_method: ApplicationMethod;
  code: string;
  /** @nullable */
  ends_at?: string | null;
  is_automatic: boolean;
  rules: PromotionRule[];
  stacking: Stacking;
  /** @nullable */
  starts_at?: string | null;
  status: PromotionStatus;
  /** @nullable */
  usage_limit?: number | null;
}