// src/lib/api/schemas/promotionStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */

export type PromotionStatus = typeof PromotionStatus[keyof typeof PromotionStatus];


export const PromotionStatus = {
  draft: 'draft',
  active: 'active',
  inactive: 'inactive',
} as const;