// src/lib/api/schemas/ruleAttribute.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */

export type RuleAttribute = 'customer_id' | 'customer_group_id' | 'order_subtotal' | 'product_id' | 'product_category_id' | 'product_brand_id' | 'variant_id' | {
  variant_property: string;
} | 'line_quantity' | 'line_unit_price';