// src/lib/api/schemas/ruleOperator.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */

export type RuleOperator = typeof RuleOperator[keyof typeof RuleOperator];


export const RuleOperator = {
  eq: 'eq',
  ne: 'ne',
  in: 'in',
  gt: 'gt',
  gte: 'gte',
  lt: 'lt',
  lte: 'lte',
} as const;